requirements-dev.txt
pytest-odoo
pydevd-odoo
